clean up CDM
