+++ great_expectations/experimental/metric_repository/metric_retriever.py
@@ -15,6 +15,7 @@ from great_expectations.experimental.metric_repository.metrics import (
     ColumnMetric,
     MetricException,
     MetricTypes,
+    TableMetric,
 )
 from great_expectations.rule_based_profiler.domain_builder import ColumnDomainBuilder
 from great_expectations.validator.exception_info import ExceptionInfo
@@ -293,3 +294,48 @@ class MetricRetriever(abc.ABC):
             if metric.metric_name == MetricTypes.TABLE_COLUMNS:
                 column_list = metric.value
         return column_list
+
+    def _get_table_row_count(self, batch_request: BatchRequest) -> Metric:
+        """Return row_count for the table.
+
+        Args:
+            batch_request (BatchRequest): For current batch.
+
+        Returns:
+            Metric: Row count for the table.
+        """
+        return self._get_table_metrics(
+            batch_request=batch_request,
+            metric_name=MetricTypes.TABLE_ROW_COUNT,
+            metric_type=TableMetric[int],
+        )
+
+    def _get_table_columns(self, batch_request: BatchRequest) -> Metric:
+        """Return column names for the table.
+
+        Args:
+            batch_request (BatchRequest): For current batch.
+
+        Returns:
+            Metric: Column names for the table.
+        """
+        return self._get_table_metrics(
+            batch_request=batch_request,
+            metric_name=MetricTypes.TABLE_COLUMNS,
+            metric_type=TableMetric[List[str]],
+        )
+
+    def _get_table_column_types(self, batch_request: BatchRequest) -> Metric:
+        """Return column types for the table.
+
+        Args:
+            batch_request (BatchRequest): For current batch.
+
+        Returns:
+            Metric: Column types for the table.
+        """
+        return self._get_table_metrics_column_types(
+            batch_request=batch_request,
+            metric_name=MetricTypes.TABLE_COLUMN_TYPES,
+            metric_type=TableMetric[List[str]],
+        )
